+++ src/CryptoNetworks.js
@@ -83,7 +83,7 @@ class CryptoNetworks {
                     "pubKeyHash": 0x30,
                     "scriptHash": 0x32,
                     "wif": 0xb0,
-                    "dustThreshold": 546
+                    "dustThreshold": 5460
                 }
             case "litecoin-testnet":
                 return {
@@ -96,7 +96,7 @@ class CryptoNetworks {
                     "pubKeyHash": 0x6f,
                     "scriptHash": 0xc4,
                     "wif": 0xef,
-                    "dustThreshold": 546
+                    "dustThreshold": 5460
                 }
             case "litecoin-regtest":
                 return {
@@ -109,7 +109,7 @@ class CryptoNetworks {
                     "pubKeyHash": 0x6f,
                     "scriptHash": 0xc4,
                     "wif": 0xef,
-                    "dustThreshold": 546
+                    "dustThreshold": 5460
                 }   
         }
     }